hg: update remotefilelog cache checking to account for size
Summary:
If a remotefilelog loose object has a size of 0, it is invalid. Since
we're already doing a syscall to check if the file exists, we might as well try
to read it's size in the process and consider it missing if the size is 0.

Previously we would report the file as there, then later when we tried to use
the file we noticed it was broken and would downlad it on demand. If many files
were broken, this adhoc one-by-one downloading would be extremely slow, so
catching it earlier so we can do batch re-downloading will speed things up.

Reviewed By: singhsrb

Differential Revision: D7046960

fbshipit-source-id: 0139f7c2cad3de617e9aae925a075bdb65f70ff5