hg: move hybrid manifest logic into it's own class
Summary:
In a future patch we're going to make the manifestlog control it's own
writes and the lifetime of its write objects. We are currently hacking up a
normal manifest to attach a treemanifest to it, which makes it awkward for it to
maintain the lifetimes of both the normal manifest and the treemanifest. Let's
move the existing logic into a class of its own so we can attach more logic in
the future.

Reviewed By: ryanmce

Differential Revision: D7083039

fbshipit-source-id: 7646b787c36af76650e77f8e6192cd266093c8ea