hg: add an option to disable filtersuspectsymlink logic
Summary:
This logic exists for the following case: if becuase of some circumstances (user error, NFS/Samba weirdness) a symlink file becomes corrupted and we check this file in while preserving the 'l' flag in the manifest, the revision becomes un-checoutable on Posix filesystems (according to the original explanations anyway).

This option basically adds an ability to take this risk. I am planning to enable this option for `ovrsource`, since it already has very few symlinks so the probability of corrupting something is pretty low.

Differential Revision: D7127333

fbshipit-source-id: f85d9f3aef676afca641280c9b7f0ecfb87b9fab