hg: add test for treemanifest and infinitepush rebundling
Summary:
There's a bug where infinitepush attempting to rebundle a bundle that
does not contain trees causes an exception because the server attempts to
prefetch those trees (which fails because there's no where to prefetch from).

This diff just adds a test for that case. The next diff will fix it.

Reviewed By: phillco

Differential Revision: D7182284

fbshipit-source-id: a3fbb576cf3318c81b18943e0f0d466aa65e54fb