Add timing information to BlobRepo upload_blob
Summary:
We want to be able to measure the time it takes to upload individual
blobs, to confirm that we don't have a concurrency issue to chase down (e.g.
blobs accidentally uploaded in series).

Measure content upload time separately, so that we know not to dive down a
rabbit hole if the measured slowness is just the time spent uploading content

Reviewed By: StanislavGlebik

Differential Revision: D7172154

fbshipit-source-id: 08729a8ffaa69a364a64f6277edfa591a8712592