join FuseChannel threads in the destructor
Summary:
Update FuseChannel to only join its worker threads in its destructor, rather
scheduling a function in the main EventBase thread to join each thread as soon
as it completes.

This makes FuseChannel a bit easier to use: the destructor can now be called at
any point in time (as long as it is not called inside one of the worker threads
themselves).  If it is called while the FuseChannel is still running this will
automatically stop the worker threads.

Reviewed By: wez

Differential Revision: D7224369

fbshipit-source-id: 4de0e7c0c677e870fd629eaaaa5ab40d134ee870