mononoke: mask the flakiness of test-infinitepush.t
Summary:
I've spent quite some time but wasn't able to get to the bottom of why
test-infinitepush.t has been flaky. For some reason error message

```
Expected Bundle2 Changegroup
```

hasn't made it's way to the hg client. Adding a multisecond delay before
sending the error "fixes" the problem, but I haven't found the actual reason
behind the problem. So since that's not a #1 priority now I suggest to postpone
fixing it for now.

Reviewed By: farnz

Differential Revision: D7246511

fbshipit-source-id: a385130e6bdc978765e04f44c6a536405ee12e02