mononoke: do not encode empty chunks in bundle2 parts
Summary:
Empty chunks results in '\x00\x00\x00\x00' being sent to the client. However
this is also a marker for end of the bundle2 part. So previously if listkeys
part were empty we were sending corrupted bundles - it had additional
'\x00\x00\x00\x00' in the end. This wasn't a problem earlier, because getbundle
was the last called wireproto command. However we want to support remotenames,
and remotenames sends listkeys wireproto request after the getbundle.

Reviewed By: farnz

Differential Revision: D7271598

fbshipit-source-id: 1304e2e74725a75726b78bcd4a9664995025191f