hg: make infinitepush rebundling use the right bundlecaps
Summary:
infinitepush rebundling wasn't passing the appropriate bundlecaps. This
caused treeonly repositories to receive flat manifests from infinitepush
bundles. This allowed treeonly clients to receive the data and convert it to
trees, but really we should just convert the data to trees on the server before
we send and not rely on the client to be able to accept flat manifests.

Now that we no longer send flat manifests, we rely on the recently added
conversion logic to turn flat manifests into trees when necessary.

This will be required for turning the server into a treeonly client anyway.

Differential Revision: D7256556

fbshipit-source-id: bbaf9d4570eebd9f200004e34044b5b1a8c53e2e